requirements/pipeline.txt
faiss-cpu
google-api-python-client
google-auth
google-auth-oauthlib
langchain
langchain_chroma
langchain-community
langchain-huggingface
langchain-core 
langchain-openai
langchain-groq
jupyterlab
matplotlib
notebook
numpy
pandas
pypdf
scikit-learn
seaborn
sentence-transformers
spacy
#fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0.tar.gz
fr-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_lg-3.7.0/fr_core_news_lg-3.7.0.tar.gz
tiktoken
ollama
openai
requests
python-dotenv
fastapi 
uvicorn
hdbscan
pypdf
umap-learn
keybert
keyphrase-vectorizers
bertopic[all]
